requirements.txt
psycopg2-binary==2.9.10
pgvector==0.3.6
sentence-transformers==3.3.1
datasets==3.2.0
numpy==1.26.4
pandas==2.2.3
matplotlib==3.9.3
seaborn==0.13.2
scipy==1.14.1
jupyter==1.1.1
python-dotenv==1.0.1
tqdm==4.67.1
